requirements.txt
tqdm==4.64.0
seaborn==0.11.0
matplotlib==3.4.3
albumentations==0.4.6
tensorflow-addons==0.11.2
sagemaker-training==4.1.0
opentsne==0.6.2
scikit-learn==1.0.2
SimpleITK==1.2.4
scikit-image==0.16.2
scipy==1.4.1
rt_utils==1.2.1
pydicom==2.1.2
dipy==1.5.0
hdbscan==0.8.28